+++ CHANGELOG.md
@@ -5,6 +5,10 @@ All notable changes to this project will be documented in this file.
 The format is mainly based on [Keep a Changelog](http://keepachangelog.com/)
 and this project adheres to [Semantic Versioning](http://semver.org/).
 
+## Unreleased
+
+* Changed: Return the certificate object in New-DomainSignedCertificate
+
 ## 2.7.0 - 2019-11-07
 
 * Added: Alias cred to the Use-VaultCredential